% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\name{RocketBase}
\alias{RocketBase}
\title{PocketBase API Client}
\description{
Instances of this class can be used for conveniently issuing
HTTP requests to remote PocketBase API instances.

Note that this client only works with admin users at the moment.
}
\examples{

## ------------------------------------------------
## Method `RocketBase$new`
## ------------------------------------------------

\dontrun{
client <- rocketbase::RocketBase$new(
    url = "https://httpbin.org",
    identity = "hebele",
    password = "hubele"
)
}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{url}}{Base URL of the remote PocketBase API instance.}

\item{\code{bare}}{Bare HTTP client for the remote PocketBase API instance.

This is a plain vanilla \code{crul::HttpClient} instance that is ready to
issue requests to remote PocketBase API instance. Indeed, this
instance should be sufficient for most API communication with
PocketBase API instances.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-RocketBase-new}{\code{RocketBase$new()}}
\item \href{#method-RocketBase-info}{\code{RocketBase$info()}}
\item \href{#method-RocketBase-print}{\code{RocketBase$print()}}
\item \href{#method-RocketBase-clone}{\code{RocketBase$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RocketBase-new"></a>}}
\if{latex}{\out{\hypertarget{method-RocketBase-new}{}}}
\subsection{Method \code{new()}}{
Creates a PocketBase API client instance.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RocketBase$new(url, identity, password)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{url}}{Base URL of the remote PocketBase API instance.}

\item{\code{identity}}{Identity credential for authentication.}

\item{\code{password}}{Password credential for authentication.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new \code{RocketBase} object.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{\dontrun{
client <- rocketbase::RocketBase$new(
    url = "https://httpbin.org",
    identity = "hebele",
    password = "hubele"
)
}
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RocketBase-info"></a>}}
\if{latex}{\out{\hypertarget{method-RocketBase-info}{}}}
\subsection{Method \code{info()}}{
Prints rudimentary information about the remote
PocketBase API instance.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RocketBase$info()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RocketBase-print"></a>}}
\if{latex}{\out{\hypertarget{method-RocketBase-print}{}}}
\subsection{Method \code{print()}}{
Provides the print function for \code{RocketBase} object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RocketBase$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RocketBase-clone"></a>}}
\if{latex}{\out{\hypertarget{method-RocketBase-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RocketBase$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
